src/test/java/ValidadorContraseñaTest.java
import org.junit.Test;

import validador_junit.ValidadorContraseña;

import static org.junit.Assert.*;

public class ValidadorContraseñaTest {
    @Test
    public void testContraseñaSegura() {
        String contraseña = "Segura123!";
        assertFalse(ValidadorContraseña.esDebil(contraseña));
    }

    @Test
    public void testContraseñaLongitudInsuficiente() {
        String contraseña = "hola1!";
        assertTrue(ValidadorContraseña.esDebil(contraseña));
    }

    @Test
    public void testContraseñaSinCaracterEspecial() {
        String contraseña = "SinCaracterEspecial123";
        assertTrue(ValidadorContraseña.esDebil(contraseña));
    }

    @Test
    public void testContraseñaComún() {
        String contraseña = "contraseña123";
        assertTrue(ValidadorContraseña.esDebil(contraseña));
    }

    @Test
    public void testContraseñaVacia() {
        String contraseña = "";
        assertTrue(ValidadorContraseña.esDebil(contraseña));
    }
}